requirements.txt
sentence-transformers>=2.2.0
chromadb>=0.4.0
pypdf2>=3.0.0
umap-learn>=0.5.3
hdbscan>=0.8.29
scikit-learn>=1.3.0
plotly>=5.15.0
streamlit>=1.22.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
tqdm>=4.0.0
